% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densefly.R
\name{BuildEmbeddedObject}
\alias{BuildEmbeddedObject}
\title{BuildEmbeddedObject}
\usage{
BuildEmbeddedObject(data, hash.length, nProj, sampling.rate,
  do.center = FALSE, seed = 19940929, old.code = FALSE)
}
\arguments{
\item{data}{an input matrix, each row is a data entry}

\item{hash.length}{the length of an input vector after one random projection, the embedding size will be hash.length*nProj}

\item{nProj}{the number of random projections}

\item{sampling.rate}{the fraction of input vector used to calculate one activation}

\item{do.center}{set this to TRUE if you hope to zero-centralize the data}

\item{seed}{a seed for generating random vectors in LSH}
}
\value{
a list of embedding results `@long` stores the long hash, `@short` stores the short hash
}
\description{
BuildEmbeddedObject
}
