% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{FindAllNearestNeighbour}
\alias{FindAllNearestNeighbour}
\title{FindAllNearestNeighbour: apply 'FindNearestNeighbour' to every rows of query}
\usage{
FindAllNearestNeighbour(query, reference, radius = 2, topN = 5,
  ncores = 1)
}
\arguments{
\item{query}{an element-by-feature query matrix}

\item{reference}{an element-by-feature reference matrix}

\item{radius}{radius for short-encoding threshold}

\item{topN}{number of nearest neighbours to return}

\item{ncores}{number of cores used in parallelization (currently not supported :(, sorry )}
}
\value{
Searching results, a dataframe including the row indices of found elements and distances
}
\description{
FindAllNearestNeighbour: apply 'FindNearestNeighbour' to every rows of query
}
