% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{FindCandidateIndex}
\alias{FindCandidateIndex}
\title{FindCandidateIndex: Find the candidates based on short encodings}
\usage{
FindCandidateIndex(query.vector, reference.matrix, minCandidate,
  radius = 1)
}
\arguments{
\item{query.vector}{a binary short-coded vector to search}

\item{reference.matrix}{a matrix of binary short-coded vector to search against}

\item{minCandidate}{minimum number of elements in the candidate list}

\item{radius}{hamming distance threshold radius}
}
\value{
the row index of the search result in the reference matrix
}
\description{
FindCandidateIndex: Find the candidates based on short encodings
}
