% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{FindNearestNeighbour}
\alias{FindNearestNeighbour}
\title{FindNearestNeighbour: Given cell's short encoding and long encoding, search its most similar cell}
\usage{
FindNearestNeighbour(query.shortEncoding, query.longEncoding,
  referenceObject, radius = 2, topN = 5, return.simple = FALSE)
}
\arguments{
\item{query.shortEncoding}{a short binary vector for multi-probe searching}

\item{query.longEncoding}{a long binary vector for searching}

\item{referenceObject}{a group of cells for reference}

\item{radius}{hamming distance threshold radius}

\item{topN}{number of nearest neighbours returned}

\item{return.simple}{whether the function returns row indices only}
}
\value{
the row name of the obtained item
}
\description{
FindNearestNeighbour: Given cell's short encoding and long encoding, search its most similar cell
}
