% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{FindNNIndex}
\alias{FindNNIndex}
\title{FindNNIndex: Find the nearest neighbour of query in the reference in hamming space}
\usage{
FindNNIndex(query.vector, reference.matrix)
}
\arguments{
\item{query.vector}{a vector to search}

\item{reference.matrix}{rows of vectors to query in}
}
\value{
the row index of the search result in the reference matrix
}
\description{
FindNNIndex: Find the nearest neighbour of query in the reference in hamming space
}
