% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densefly.R
\name{GetActivations}
\alias{GetActivations}
\title{Convert an input vector into activations}
\usage{
GetActivations(data, embedding.size, sampling.rate)
}
\arguments{
\item{data}{an input matrix, each row is a data entry}

\item{embedding.size}{length of embedded vector, i.e. number of columns after embedding}

\item{sampling.rate}{the fraction of input vector used to calculate one activation}
}
\value{
a matrix with the same number of rows as the input, each row is an embedded vector
}
\description{
Convert an input vector into activations
}
