% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densefly.R
\name{GetLongHash}
\alias{GetLongHash}
\title{GetLongHash}
\usage{
GetLongHash(activations)
}
\arguments{
\item{activations}{an activation matrix, each row represents a data entry's activations}
}
\value{
a binary matrix, each row is an embeded data entry
}
\description{
GetLongHash
}
