% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densefly.R
\name{GetShortHash}
\alias{GetShortHash}
\title{GetShortHash}
\usage{
GetShortHash(activations, nProj)
}
\arguments{
\item{activations}{an activation matrix, each row represents a data entry's activations}

\item{nProj}{the number of random projections}
}
\value{
a binary matrix, each row is an embeded data entry
}
\description{
GetShortHash
}
