% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashset.R
\name{HashsetInsert}
\alias{HashsetInsert}
\title{HashsetInsert: insert a vector element into the hashset}
\usage{
HashsetInsert(vector, hash.table)
}
\arguments{
\item{vector}{vector to insert}

\item{hash.table}{a hash table object created using package `hash`}
}
\value{
The function should change the hash table in the parameter and has no output
}
\description{
HashsetInsert: insert a vector element into the hashset
}
