% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashset.R
\name{HashsetInsertMatrix}
\alias{HashsetInsertMatrix}
\title{HashsetInsertMatrix: insert a matrix (a set of vector elements, each row is a vector) into the hashset}
\usage{
HashsetInsertMatrix(matrix, hash.table)
}
\arguments{
\item{matrix}{rows of vectors to insert}

\item{hash.table}{a hash table object created using package `hash`}
}
\value{
a hash table object created using package `hash`
}
\description{
HashsetInsertMatrix: insert a matrix (a set of vector elements, each row is a vector) into the hashset
}
