% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashset.R
\name{HashsetQuery}
\alias{HashsetQuery}
\title{HashsetQuery: query a vector element in the hashset}
\usage{
HashsetQuery(vector, hash.table)
}
\arguments{
\item{vector}{vector to query}

\item{hash.table}{a hash table object created using package `hash`}
}
\value{
a boolean variable indicating if the query exists in the hashset
}
\description{
HashsetQuery: query a vector element in the hashset
}
