% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashset.R
\name{HashsetQueryMatrix}
\alias{HashsetQueryMatrix}
\title{HashsetQuery: query a matrix (a set of vector elements, each row is a vector) in the hashset}
\usage{
HashsetQueryMatrix(matrix, hash.table)
}
\arguments{
\item{matrix}{rows of vectors to query}

\item{hash.table}{a hash table object created using package `hash`}
}
\value{
a boolean vector indicating if each row exists in the hashset
}
\description{
HashsetQuery: query a matrix (a set of vector elements, each row is a vector) in the hashset
}
