% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{Hashtable.Key}
\alias{Hashtable.Key}
\title{Hashtable.key: convert a vector into hashtable key}
\usage{
Hashtable.Key(vector)
}
\arguments{
\item{vector}{a numeric vector}
}
\value{
a character variable
}
\description{
Hashtable.key: convert a vector into hashtable key
}
