% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtable.R
\name{HashtableInsertMatrix}
\alias{HashtableInsertMatrix}
\title{HashtableInsertMatrix: insert a matrix (a set of vector elements, each row is a vector) into the hashtable}
\usage{
HashtableInsertMatrix(matrix, hash.table)
}
\arguments{
\item{matrix}{rows of vectors to insert}

\item{hash.table}{a hash table object created using package `hash`}
}
\value{
a hash table object created using package `hash`
}
\description{
HashtableInsertMatrix: insert a matrix (a set of vector elements, each row is a vector) into the hashtable
}
\examples{
# create sample data
x <- matrix(
    c(1,2,5,1,3,1,
      1,5,3,2,4,5,
      3,4,2,1,1,0),nrow=3,byrow=TRUE)
rownames(x)<- as.character(1:3)
h<- hash()
h<-HashtableInsertMatrix(matrix=x, hash.table=h)
h
# it should return:
<hash> containing 3 key-value pair(s).
125131 : 1
153245 : 2
342110 : 3
}
