% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{MappingCells}
\alias{MappingCells}
\title{MappingCells: apply query operations on the query matrix}
\usage{
MappingCells(query, reference, reference.label, cores)
}
\arguments{
\item{query}{query matrix}

\item{reference}{reference matrix}

\item{reference.label}{labels of the reference}

\item{cores}{number of cores used}
}
\description{
MappingCells: apply query operations on the query matrix
}
