% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist-data.R
\docType{data}
\name{mnist.test.x}
\alias{mnist.test.x}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class `"cross"`; see [qtl::read.cross()].}
\usage{
data(mnist.test.x)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(mnist.test.x)
data(mnist.test.y)
data(mnist.train.x)
data(mnist.train.y)


}
\references{
Yann LeCun, Corinna Cortes, Christopher J.C. Burges,
THE MNIST DATABASE of handwritten digits
([MNIST](http://yann.lecun.com/exdb/mnist/))
}
\keyword{datasets}
