% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{query.one.cell}
\alias{query.one.cell}
\title{query.one.cell: a wrapper for query operations}
\usage{
query.one.cell(idx)
}
\arguments{
\item{idx}{row index of cell in the query matrix}
}
\description{
query.one.cell: a wrapper for query operations
}
