% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{SearchCell}
\alias{SearchCell}
\title{SearchCell: Given cell's short encoding and long encoding, search its most similar cell}
\usage{
SearchCell(query.shortEncoding, query.longEncoding, referenceObject)
}
\arguments{
\item{query.shortEncoding}{a short binary vector for multi-probe searching}

\item{query.longEncoding}{a long binary vector for searching}

\item{referenceObject}{a group of cells for reference}
}
\value{
the row name of the obtained cell
}
\description{
SearchCell: Given cell's short encoding and long encoding, search its most similar cell
}
