% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densefly.R
\name{GetShortHash}
\alias{GetShortHash}
\title{GetShortHash}
\usage{
GetShortHash(activations, nProj)
}
\arguments{
\item{activations}{an activation matrix, each row represents a data entry's activations}

\item{nProj}{the number of random projections}
}
\value{
a binary matrix, each row is an embeded data entry
}
\description{
GetShortHash
}
\examples{
activations<-c(-3, -2, 1, 0, 5,   -1, 2, 1, 0, 7,   0, -2, 0, 0, 1)
activations<-as.matrix(t(activations))
GetShortHash(activations,5) # it is supposed to be a matrix of (1,1,0)
}
