% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{MappingCells}
\alias{MappingCells}
\title{Title}
\usage{
MappingCells(query, reference, reference.label, cores)
}
\arguments{
\item{query}{an embedded query object}

\item{reference}{an embedded reference object}

\item{reference.label}{the reference dataset's label}

\item{cores}{number of cores used, =1 for serial execution}
}
\value{
labels transfered from reference after mapping
}
\description{
Title
}
\examples{
# A MNIST example
## Reading mnist datasets
library("dslabs")
mnist <- read_mnist()

mnist.train.x <- mnist$train$images
mnist.train.y <- mnist$train$labels
rownames(mnist.train.x)<-1:nrow(mnist.train.x)

mnist.test.x <- mnist$test$images
mnist.test.y <- mnist$test$labels
rownames(mnist.test.x)<-1:nrow(mnist.test.x)


## Building embedded object
set.seed(19940929)
reference <- BuildEmbeddedObject(data=mnist.train.x,
                                 nProj=125, hash.length=8,
                                 sampling.rate=0.3, do.center=TRUE)
set.seed(19940929)
query     <- BuildEmbeddedObject(data=mnist.test.x,
                                 nProj=125, hash.length=8,
                                 sampling.rate=0.3, do.center=TRUE)

## Mapping
tic()
mapped<-mapping.cells(query=query, reference=reference, reference.label=mnist.train.y,cores=12)
toc()
### 121.387 sec elapsed


## Evaluation
kappa2(matrix(cbind(mapped,mnist.test.y), ncol=2, byrow=FALSE))$value
### 0.909291617556118
}
