% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{query.one.cell}
\alias{query.one.cell}
\title{query.one.cell A function wrapper for query dataset mapping}
\usage{
query.one.cell(idx)
}
\arguments{
\item{idx}{the row index of data matrix}
}
\value{
the index of nn data item
}
\description{
query.one.cell A function wrapper for query dataset mapping
}
